// constants/enums/notification-event.ts
export enum NotificationType {
  SERVICE_REQUEST = 'Service Request',
  NOTICE_AGREEMENT = 'Notice Agreement',
  RENT_CREATED = 'Rent Created',
  USER_ADDED_TO_PROPERTY = 'User Added to Property',
  USER_SIGNED_UP = 'User Signed Up',
  LEASE_SIGNED = 'Lease Signed',
  PROPERTY_CREATED = 'Property Created',
}